Add arm64 architecture to Darkest project file and fix compiler warnings.
git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/darkest_before_dawn@14782 afe8dcc5-9094-451b-8773-5f495042a737

+++ ios/Darkest/OpenGLController.m
@@ -120,7 +120,7 @@
         
         CGPoint pt = [m_arrTouches[i] locationInView:self.view];
         [self RecalcTouchPosForCGE:&pt];
-        CGEA_Motion(pt.x, pt.y, i);
+        CGEA_Motion(pt.x, pt.y, (int)i);
     }
     
     CGEA_Update();
@@ -170,7 +170,7 @@
         
         CGPoint pt = [touch locationInView:self.view];
         [self RecalcTouchPosForCGE:&pt];
-        CGEA_TouchDown(pt.x, pt.y, nFingerIdx);
+        CGEA_TouchDown(pt.x, pt.y, (int)nFingerIdx);
     }
     
     [super touchesBegan:touches withEvent:event];
@@ -188,7 +188,7 @@
         
         CGPoint pt = [touch locationInView:self.view];
         [self RecalcTouchPosForCGE:&pt];
-        CGEA_TouchUp(pt.x, pt.y, nFingerIdx);
+        CGEA_TouchUp(pt.x, pt.y, (int)nFingerIdx);
     }
 
     [super touchesEnded:touches withEvent:event];
@@ -206,7 +206,7 @@
         
         CGPoint pt = [touch locationInView:self.view];
         [self RecalcTouchPosForCGE:&pt];
-        CGEA_TouchUp(pt.x, pt.y, nFingerIdx);
+        CGEA_TouchUp(pt.x, pt.y, (int)nFingerIdx);
     }
     
     [super touchesCancelled:touches withEvent:event];
